Added Mediocre Tic-Tac-Toe game and Router
I created a tic-tac-toe game that is not finished. Its components are within components/games/TicTacToe. The game is play able but
the styling is very bad and I want to change some of the logic to make it
more flexible and robust, since I'm planning on adding AI and online
features in the soon future.
Besides the tic-tac-toe game, I also added two pages to the pages folder
and added routes so the user can get to the tic-tac game from the home
page. The home page
might be the page that gives the users game options and redirects them
to the appropiate game page.